requirements.txt
streamlit
pandas
snowflake-connector-python
plotly
scikit-learn 
pandas
joblib
uvicorn 
fastapi
mlflow